refactor(components): migrate Select component to TypeScript
feature/typescript-select

feature/typescript-select

feature/typescript-select
