+++ src/components/Select/Select.docs.mdx
@@ -1,5 +1,5 @@
 import { Status, Props, Story } from '../../../.storybook/components';
-import Select from './Select';
+import { Select } from './Select';
 
 # Select
 
@@ -9,7 +9,6 @@ A select dropdown is a form field that allows users to choose one of the multipl
 An example being the expiration date for a users credit card.
 
 <Story id="forms-select--base" />
-
 <Props of={Select} />
 
 ## When to use it
