+++ src/components/Select/index.ts
@@ -13,6 +13,6 @@
  * limitations under the License.
  */
 
-import Select from './Select';
+import { Select } from './Select';
 
 export default Select;
